+++ lib/domain/services/notification_service.dart
@@ -1,0 +1,3 @@
+abstract class NotificationService {
+  Future<void> sendNotification(String notId, String appToken);
+}