+++ lib/domain/services/subscription_service.dart
@@ -1,0 +1,5 @@
+import '../../data/model/inngage/subscription_request.dart';
+
+abstract class SubscriptionService {
+  Future<void> subscription(SubscriptionRequest subscription);
+}